add vine/ivy language configuration to vscode plugin
