+++ lsp/client/src/ivy-config.json
@@ -1,0 +1,23 @@
+{
+  "comments": {
+    "lineComment": "//",
+    "blockComment": ["/*", "*/"]
+  },
+  "brackets": [
+    ["{", "}"],
+    ["(", ")"]
+  ],
+  "colorizedBracketPairs": [],
+  "autoClosingPairs": [
+    { "open": "{", "close": "}" },
+    { "open": "(", "close": ")" }
+  ],
+  "surroundingPairs": [
+    ["(", ")"],
+    ["{", "}"]
+  ],
+  "wordPattern": {
+    "pattern": "(::\\p{ID_Continue}+)+|\\p{ID_Start}\\p{ID_Continue}*",
+    "flags": "ug"
+  }
+}
