+++ lsp/client/src/vine-grammar.json
@@ -1,0 +1,9 @@
+{
+  "name": "rustdoc",
+  "patterns": [
+    {
+      "include": "source.rust"
+    }
+  ],
+  "scopeName": "source.vine"
+}
