+++ lsp/client/tsconfig.json
@@ -4,7 +4,8 @@
     "target": "es2023",
     "lib": ["es2023"],
     "outDir": "dist",
-    "rootDir": "src"
+    "rootDir": "src",
+    "checkJs": true
   },
   "include": ["src"]
 }
